==> 04b-recap_donnees_dispo.qmd <==
# Récapitulatif des données disponibles {#sec-recap_donnees}

## Discussion sur la construction, portée et limites de ces données

```{=html}
<iframe src="TVM2022_PresentationDesDonnees.pdf" width="100%" height="400px"></iframe>
```

[Cliquer ici pour télécharger la présentation](TVM2022_PresentationDesDonnees.pdf).

Un point crucial à retenir est l'importance de toujours accompagner l'analyse d'une discussion critique des données mobilisées : quelles définitions sont utilisées pour établir les catégories que j'utilise (dans notre exemple, quelle définition de la forêt a été retenue) ? Y a-t-il des différences ou des nuances entre ce qui est mesuré et le phénomène que je veux analyser (dans notre exemple, la variation du couvert forestier n'est pas à proprement parler de la déforestation) ? Quels problèmes de qualité affectent mes données (dans notre exemple, les erreurs ou approximations qui affectent les données WDPA) ? etc.

## Liste des variables

On réalise un résumé des données disponible. Les fonctions `str()` et `summary()` sont les plus couramment utilisées pour avoir une idée rapide du contenu des tables pendant qu'on étudie les données. Le package {} facilite l'élaboration d'une description rapide d'un jeu de données. C'est l'exemple qu'on prend ci-desous.

> TODO: le faire avec un package qui produit un joli rendu.

```{r}
load("data/ch3_AP_Vahatra.rds")
str(AP_Vahatra)
```

### Exercices

Produire le même tableau de résumé pour les autres tables :

## Autres sources d'information

-   [DHS](https://dhsprogram.com/data/available-datasets.cfm) et [MICS](http://mics.unicef.org/) : principales enquêtes socio-sanitaires
-   [IHSN](http://www.ihsn.org/) : la plupart des enquêtes nationales
-   [World Bank Microdata catalog](http://microdata.worldbank.org/index.php/catalog/central) : microdonnées de nombreuses autres enquêtes (notamment des grosses enquêtes sur l'eau par la Banque mondiale)
-   [WAPOR](https://wapor.apps.fao.org/home/1) : données satellitaires préprocessées de productivité : rapport biomass water productivity, évapotranspiration, précipitations... (cliquez sur l'onglet « layers » en bas à gauche de l'écran pour voir la liste des donnés disponibles)

Il y a encore plein d'autres choses sur l'accès à l'eau dans les établissements de santé ou les écoles, ou encore le satellitaire, mais c'est plus diffus et je ne veux pas vous noyer. On prépare un outil en ligne avec la KfW qui permettra aux REP et partenaires de mieux s'y retrouver.
