==> 06-matching_AP.qmd <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

# Méthode d'appariement

```{=html}
<iframe src="TVM_Matching.pdf" width="100%" height="400px"></iframe>
```

[Cliquer ici pour télécharger la présentation](TVM_Matching.pdf).

On a vu dans le chapitre précédent que les comparaisons simples réalisées entre les premières et les dernières aires à avoir été formellement protégées pose problème.

On va maintenant chercher à renforcer la comparabilité entre le goupe de traitment et le groupe de contrôle en réalisant un appariemment (cf. diapos de présentation).

On va utiliser le package {MatchIt}: ne pas hésiter à se référer à la documentation du package : \[TODO: insérer le lien vers la doc\]

On va commencer par réaliser quelques ajustements, car {MatchIt} requiert qu'aucune valeur des variables mobilisées ne soit manquante. On va donc retirer les observations comportant des `NA`.

```{r}
library(tidyverse) # Simplifie la manipulation de données
library(lubridate) # Simplifie les opérations sur des dates
library(sf) # Pour traiter les données spatiales
library(MatchIt) # Pour réaliser les appariements.
library(cobalt) # Pour les tests d'équilibre sur l'appariement
library(gt) # Pour faire de jolies tables
library(stargazer) # Pour préssenter les résultats de régressions

# Désactiver les notations scientifiques
options(scipen =999)
# On recharge les données préparées dans le chapitre 3
load("data/ch3_AP_Vahatra.rds")

# Harmoniser les données entre avant et après.
load("data/rct_AP_Mada.rds")


rct_AP_Mada_noNA <- rct_AP_Mada %>%
  # On enlève les observations pour lesquelles il manque des valeurs
  filter(!is.na(`Déforestation 1996-2016 (%)`)) %>%
  filter(!is.na(dist_ville)) %>%
  # La vatiable de traitement doit être recodée en [0, 1]
  mutate(traitement = ifelse(Groupe == "Traitement", 1, 0)) %>%
  rename(surface_ha = `Surface (ha)`, 
         couv_foret_96 = `Couvert forestier en 1996 (%)`)

rct_AP_Mada_noNA %>%
  group_by(Groupe) %>%
  summarize(`Nombre d'aires protégées` = n()) %>%
  gt() %>%
  tab_header("Observations par groupe avant appariemment") %>%
  tab_source_note("Source : Association Vahatra et Carvalho et al. 2018")
```

Pour commencer, on va estimer le un modèle qui estime dans quel mesure la propension pour une aire d'avoir été protégée avant 2015 dépend de sa taille, de son taux de couverture forestière en 1996, de son altitude, de son caractère accidenté et de sa distance d'une ville d'au moins 5000 habitants.

Cette spécification peut se représenter selon l'équation suivante : \[TODO: revoir la spécification au format standard\]

$$
T = \alpha + \beta_{1}A + \beta_{2}B + \beta_{3}C + \beta_{4}D + \beta_{5}E + \varepsilon
$$ {#eq-psm}

Où Y est le traitement, A est la taille (surface en hectares, B le taux de couverture forestière en 1996, C l'altitude, D, le caractère accidenté et E le temps de parcours à une ville d'au moins 5000 habitants.

Cette même formule est encodée en R de la manière suivante :

```{r}
#| code-fold: false
pscor <- traitement ~  surface_ha + 
                       couv_foret_96 + 
                       altitude +
                       indice_accidente + 
                       dist_ville
```

On va maintenant réaliser une régression pour connaître l'influence de ces facteurs dans la désignation des aires comme protégées.

```{r}
#| output: false
reg_select <- glm(formula = pscor,
                  family = binomial(link = "probit"),
                  data = rct_AP_Mada_noNA)

match_out1 <- stargazer(reg_select, type = "html") %>%
  str_replace_all("\\*", "\\\\*") 

# Dans un bloc plus bas et non affiché, on a le code suivant
# cat(match_out1)
```
```{r}
#| output: asis
#| echo: false

cat(match_out1)
```


On va maintenant utiliser ce modèle pour comparer les aires protégées traitées en premier par rapport à celles traitées plus récemment.

```{r}
# Calcul du matching
def_96_16_match <- matchit(formula = pscor,
                           family = binomial(link = "probit"),
                           method = "nearest",
                           discard = "both",
                           replace = FALSE,
                           distance = "glm",
                           data = rct_AP_Mada_noNA)

print(def_96_16_match)
```

On peut maintenant observer les équilibres entre les groupes traités et contrôle avant et après l'appariement.

```{r}
summary(def_96_16_match)
```

> **Exercice** : Etudiez les tables ci-dessus. Quel effet a eu l'appariement sur l'équilibre des variables entre le groupe de traitement et le groupe de contrôle ? Combien d'observation ont été écartées.

On peut observer la distance entre groupe de traitement et de contrôle.

```{r}
plot(def_96_16_match, type = "jitter", interactive = FALSE)
```

On peut également représenter l'équilibre entre les variables avant et après traitement avec les graphiques suivants.

```{r}
bal.plot(def_96_16_match, var.name = "dist_ville", which = "both")
```

> **Exercice** : Quel effet a eu l'appariement sur la varialbe de distance à la ville ? Les autres variables d'appariement produisent-elles un effet aussi visible ?

Le modèle qu'on utilise pour estimer l'impact est très proche de celui exposé ci-dessus, à la différence que la variable de traitement passe dans la partie droite, et qu'elle est remplacée par la déforestation.

$$
Y = \alpha + \beta_{0}T + \beta_{1}A + \beta_{2}B + \beta_{3}C + \beta_{4}D + \beta_{5}E + \varepsilon
$$

Où Y est la déforestation, T est le traitement, A est la taille (surface en hectares, B le taux de couverture forestière en 1996, C l'altitude, D, le caractère accidenté et E le temps de parcours à une ville d'au moins 5000 habitants.

Cette formule est codée en R de la manière suivante :

```{r}
# On extrait la donnée de l'appariement

#| code-fold: false
estimp <- `Déforestation 1996-2016 (%)` ~   
                          traitement +
                          surface_ha + 
                          couv_foret_96 + 
                          altitude +
                          indice_accidente + 
                          dist_ville
```

On va donc réaliser une régression, en tenant compte des pondérations générées par l'algorithme d'appariement (variable "weight").

```{r}
#| output: false

# On extrait les données de l'appariement
def_96_16_match_data <- match.data(def_96_16_match)
# On effectue une régression simple avec la formule précédente
def_96_16_match_est <- lm(formula = estimp,
                          data = def_96_16_match_data,
                          weights = weights)
# On visualise les résultats
match_out2 <- stargazer(def_96_16_match_est, type = "html") %>%
  str_replace_all("\\*", "\\\\*") 

# Pour afficher le rendu (dans un bloc juste après et pas affiché)
# cat(match_out2)
```
```{r}
#| output: asis
#| echo: false

cat(match_out2)
```


### Exercice simple

Analysez, interprétez et critiquez les résultats ci-dessus.

### Exercice intermédiaire

Ajoutez des variables d'interne et modifiez les paramètres de la fonction de matching.

### Exercice avancé

Réalisez une analyse analogue avec les données de déforestation TMF. Rédigez une analyse interprétative.
